required changes move functions to service layer
